requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.2
boto3==1.34.29
botocore==1.34.29
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
cfscrape==2.1.1
charset-normalizer==3.3.2
cramjam==2.8.1
cryptography==42.0.0
distlib==0.3.8
exceptiongroup==1.2.0
fake-useragent==1.4.0
fastparquet==2024.2.0
filelock==3.13.1
flake8==6.0.0
fsspec==2024.2.0
h11==0.14.0
identify==2.5.33
idna==3.6
importlib-resources==6.1.1
jmespath==1.0.1
mccabe==0.7.0
nodeenv==1.8.0
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.3
platformdirs==4.1.0
pre-commit==3.1.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
s3transfer==0.10.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
virtualenv==20.25.0
wsproto==1.2.0
zipp==3.17.0
